# Optimizing the BFS preloading prefix computation