--- a/akd/src/utils.rs
+++ b/akd/src/utils.rs
@@ -9,37 +9,7 @@
 // 2. For each node in current_nodes set, check if each child is in prefix hashmap
 // 3. If so, add child label to batch set
 
-use crate::{NodeLabel, EMPTY_LABEL, EMPTY_VALUE};
-use std::collections::HashSet;
-
-// Builds a set of all prefixes of the input labels
-pub(crate) fn build_prefixes_set(labels: &[NodeLabel]) -> HashSet<NodeLabel> {
-    let mut prefixes_set = HashSet::new();
-    for label in labels {
-        for len in 0..(label.get_len() + 1) {
-            prefixes_set.insert(label.get_prefix(len));
-        }
-    }
-    prefixes_set
-}
-
-pub(crate) fn build_lookup_prefixes_set(labels: &[NodeLabel]) -> HashSet<NodeLabel> {
-    let mut lookup_prefixes_set = HashSet::new();
-    for label in labels {
-        // We need the actual node for lookup too
-        lookup_prefixes_set.insert(*label);
-        for len in 0..(label.get_len() + 1) {
-            // Sibling prefixes unfortunately do not cover all the nodes we will need for
-            // a lookup proof. Although we can figure out which nodes are exactly needed
-            // this will require basically doing a pre-lookup.
-            // Instead here we load the prefixes as well.
-            // This combination (sibling- + self-prefixes) covers all the nodes we need.
-            lookup_prefixes_set.insert(label.get_prefix(len));
-            lookup_prefixes_set.insert(label.get_sibling_prefix(len));
-        }
-    }
-    lookup_prefixes_set
-}
+use crate::{EMPTY_LABEL, EMPTY_VALUE};
 
 pub(crate) fn empty_node_hash() -> crate::Digest {
     crate::hash::merge(&[crate::hash::hash(&EMPTY_VALUE), EMPTY_LABEL.hash()])
@@ -48,7 +18,7 @@ pub(crate) fn empty_node_hash() -> crate::Digest {
 // Creates a byte array of 32 bytes from a u64
 // Note that this representation is big-endian, and
 // places the bits to the front of the output byte_array.
-#[allow(dead_code)]
+#[cfg(any(test, feature = "public-tests"))]
 pub(crate) fn byte_arr_from_u64(input_int: u64) -> [u8; 32] {
     let mut output_arr = [0u8; 32];
     let input_arr = input_int.to_be_bytes();
@@ -56,7 +26,7 @@ pub(crate) fn byte_arr_from_u64(input_int: u64) -> [u8; 32] {
     output_arr
 }
 
-#[allow(dead_code)]
+#[allow(unused)]
 #[cfg(any(test, feature = "public-tests"))]
 pub(crate) fn random_label(rng: &mut rand::rngs::OsRng) -> crate::NodeLabel {
     use crate::rand::Rng;