Previously we computed the preloading of nodes by first taking the insertion set of labels, computing all prefixes, and then loading the nodes in BFS manner and checking each node's label against the prefix set. For N labels to insert, this was a set of size `256 * N`, which resulted in about `256 * N` work.

Now, we are using a BinaryHeap to insert the labels in sorted order (sorted by label value), and then instead of building a prefix set, we do a binary search to check if the current node during the BFS traversal is contained as a prefix in the sorted insertion set of labels. This is roughly `N * log(N)` work to build the sorted list, and then about `log^2(N)` work for prefix-checking during insertion. This should be much faster -- see benchmarks below.

Before this change:
Inserting 100k nodes into an empty tree:
```
[00:02:50.423] INFO   Preload of tree (1 objects loaded), took 31.261086083 s (append_only_zks:217)
```
Inserting another 100k nodes into the tree subsequently:
```
[00:06:14.654] INFO   Preload of tree (150813 objects loaded), took 34.730958 s (append_only_zks:217)
```

After this change:

Inserting 100k nodes into an empty tree:
```
[00:02:18.657] INFO   Preload of tree (1 objects loaded), took 0.005568666 s (append_only_zks:265)
```
Inserting another 100k nodes into the tree subsequently:
```
[00:05:06.851] INFO   Preload of tree (151035 objects loaded), took 1.7127710409999999 s (append_only_zks:265)
```

-----------

**Edit**: Those above numbers are run with `--profile = dev` mode. In `--profile = release` mode, the numbers are less drastic but about the same in terms of scaling:

Before:
```
[00:00:21.253] INFO   Preload of tree (1 objects loaded), took 4.403417458 s (append_only_zks:217)
Batch insertion (100000 initial leaves) took 22996 ms

[00:00:45.232] INFO   Preload of tree (150755 objects loaded), took 4.546898041 s (append_only_zks:217)
Batch insertion (100000 initial leaves, 100000 inserted leaves) took 25390 ms
```

After:
```
[00:00:16.759] INFO   Preload of tree (1 objects loaded), took 0.000585958 s (append_only_zks:265)
Batch insertion (100000 initial leaves) took 17196 ms

[00:00:34.034] INFO   Preload of tree (151191 objects loaded), took 0.199543458 s (append_only_zks:265)
Batch insertion (100000 initial leaves, 100000 inserted leaves) took 19388 ms
```