@slawlor 2022-12-26
OK so the benchmarking here is against an in-memory DB right? So we're only measuring the computational work, not any of the actual DB timing, correct?