@kevinlewi 2022-12-26
Yep, that's correct, these benchmarks were against an in-memory DB.

The actual loading of nodes from the DB in the before and after should be the same -- we aren't changing anything there, since we are still loading the same # of nodes and in the same manner. The only thing that is changing is a more efficient way to compute which nodes we should be loading in.